% Generated by roxygen2 (4.0.0.99): do not edit by hand
\docType{data}
\name{lrlake}
\alias{lrlake}
\title{Little Rock Lake zooplankton dataset}
\format{A data frame with 592 observations on the following 18 variables.}
\source{
http://lter.limnology.wisc.edu/
}
\description{
Contains time series for 10 dominant crustaceous species of zooplanking
sampled from Little Rock Lake, Wisconsin. Samples come from two basins: one
treated to lower pH and the other an untreated reference.
}
\examples{
data(lrlake)
x = subset(lrlake, Basin == "Reference", LRL.Day)
y = subset(lrlake, Basin == "Reference", -(1:8))
matplot(x, y, type = "l", lty = 1)
}
\references{
Keitt, T. H. 2008. Coherent ecological dynamics induced by
large-scale disturbance. Nature 454:331-4. doi:10.1038/nature06935.
}
\keyword{datasets}

